Added menu option and back to top button
When someone access the page, he/she will take time to scroll down until the section he/she wants. With menu option which can click and direct to the section user wants, user can save a lot of time. The back to top button can let the users back to the top of this page and select the next section they want to browse from the menu.