Added CSS for menu and back to top button
When clicking on the list of menu, the page will direct user to the section selected, the CSS added allow the process of scrolling down run smoothly. The CSS for back to top button makes the button disappear when the page is at the top and appear when the page is not at the top.